update review.java with total sentiment
This finds the sentiment value of each word, excluding punctuation which does not add to the value. Letters are added together until the method gets to a space, and then a method is called that finds the sentiment value of the word in its entirety. The value is stored, and then returned with return totalSentiment.