added starRating()
starRating() looks at the total sentiment value of the review and assigns a number of stars based on this value. The "if" statement allows for the sentiment value to be converted into a number of stars. So, "if" the sentiment value of the review is less than zero, the star rating will be 1 or 2 stars. starRating() returns the rating between 1 and 5 stars.