Final totalSentiment
I created the totalSentiment method, which returns senValue, the sentiment value of a review. This method works because it finds the sentiment value of each word, without regard to punctuation. The result of this method assigns a sentiment value to a review. 

@bdurling3
@xaamaro

Is my indentation correct? How does my method differ from yours? Is there any way to condense my method down into fewer lines of code?