final starRating()
This method returns an integer from 1 to five which corresponds to the sentiment of the review. The if/else if/else statements allow the value of totalSentiment to be correctly matched with the amount of stars. 

@sevandersip
@parkertopak

Is this method more/less efficient than yours, and if it is less efficient, how could I condense it?